Change CQRS service registration to scoped lifetime
There was an issue when running in the context of ASP.NET Core:
```
System.InvalidOperationException HResult=0x80131509 Message=Cannot resolve scoped service 'CodeWinden.CQRS.Proxies.QueryHandlerProxy2[CodeWinden.CQRS.IQueryHandler1[{ResultClass}], {ResultClass}]' from root provider. 
```

**Reasoning**:
The `IServiceProvider` injected into `CQRSService` is the root service provider because `CQRSService` is a singleton. Scoped services can only be resolved from a scoped service provider (like the one created per HTTP request in ASP.NET Core), not from the root provider.

So we need to update it to Scoped to be able to work properly. (Since it's a very lightweight-class this shouldn't have any significant impact on performance and/or memory)